@@ -27,21 +27,21 @@
   },
   "homepage": "https://github.com/bakerstreet-industries/ts-log-class#readme",
   "devDependencies": {
-    "@types/chai": "^4.0.6",
+    "@types/chai": "^4.1.4",
     "@types/chai-spies": "0.0.0",
-    "@types/mocha": "^2.2.44",
+    "@types/mocha": "^2.2.48",
     "chai": "^4.1.2",
     "chai-spies-next": "^0.9.3",
-    "coveralls": "^3.0.0",
-    "mocha": "^4.0.1",
+    "coveralls": "^3.0.2",
+    "mocha": "^4.1.0",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^11.3.0",
-    "publish-please": "^2.3.1",
-    "ts-loader": "^3.2.0",
+    "nyc": "^11.9.0",
+    "publish-please": "^3.2.0",
+    "ts-loader": "^3.5.0",
     "ts-node": "^3.3.0",
-    "tslint": "^5.8.0",
-    "typescript": "^2.6.2",
-    "webpack": "^3.9.1"
+    "tslint": "^5.11.0",
+    "typescript": "^2.9.2",
+    "webpack": "^3.12.0"
   },
   "nyc": {
     "include": [
@@ -61,7 +61,9 @@
     "instrument": true
   },
   "dependencies": {
-    "circular-json": "^0.5.1"
+    "circular-json": "^0.5.5",
+    "logdown": "^3.2.3",
+    "raven-js": "^3.26.4"
   },
   "typings": ".webpack/src/main.d.ts"
 }
